test: replace .jsx snapshot file with .tsx
